use dspace.url config for base
Rather than extracting the base url from the HTTP Request, just always
use the configured dspace.url. There can be conditions where it gets
generated in a way that is incompatible with the url that is registered
with GlobusAuth. This assumes that dspace.url is always set to the same
URL registered with GlobusAuth.