Fix substitution of error messages on storage config errors
There was supposed to be a substitution of error messages in the case
that storage configuration failed on a collection's endpoint, but
apparently the substitution was never actually performed. That
resulted in error messages to the user where $publishUser was
displayed. Fix that though it assumes that the username property is
available in globus-auth.cfg.