Fix the edit dataset page
The edit dataset page uses underscores as separators, but some of our
metadata field definitions also use underscores in the name, so the
conflict prevents the update from occurring. This changes so that the
internal field separators are done with a double underscore (__) and
the corresponding split/concatenate changes are made.