More null errors in processing queries
It seems that perhaps the root cause is that DSpace assumes that all
query requests will be coming from clients that have a logged in
session and thus parts of the query building can get state from the
session. When non-session clients (like a web crawler or simple curl)
hit some of the query generation links, incomplete queries get built
and processing those queries results in errors.

This is an attempt to find various places where the query for
facets/filters is getting generated and put protective checks around
them. Hopefully, this gets them all.