Change URLs for callouts to the Globus WebApp
With updates to the Globus WebApp, various callout locations have
changed. This updates to be consistent with the updated Globus WebApp
URLs.

Also did some refactoring to bring the WebApp Integration specific
functions out of the meta-Globus.java file and into a new file:
GlobusWebAppIntegration.java (which is a mouth-full).